+++ packages/types/src/global.ts
@@ -2,9 +2,10 @@
 // SPDX-License-Identifier: GPL-3.0
 
 import {Provider} from '@ethersproject/abstract-provider';
+import {JsonRpcProvider} from '@ethersproject/providers';
 import '@subql/types-core/dist/global';
 
 declare global {
   const api: Provider;
-  const unsafeApi: Provider | undefined;
+  const unsafeApi: JsonRpcProvider | undefined;
 }