+++ packages/types/CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Type for unsafeApi to from Provider to JsonRpcProvider (#291)
 
 ## [3.6.0] - 2024-04-30
 ### Changed